work on the html and css a little
