+++ 02week/ticTacToe.css
@@ -1,0 +1,19 @@
+body {
+  background-color: bisque;
+  margin: 0;
+  width: 100%;
+  height: 100%;
+}
+
+/* h1 {
+  need to assign a font
+} */
+
+#board { 
+  width: 75%;
+  margin: 15%;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-rows: repeat(3, 1fr);
+  height: 60%;
+}
