+++ 02week/ticTacToe.html
@@ -1,0 +1,25 @@
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="../../../ACA_web101_Natalie_Kendrick/reset.css">
+  <link rel="stylesheet" href="ticTacToe.css">
+  <title>Tic Tac Toe!</title>
+</head>
+<body>
+  <main>
+    <h1>Welcome to Tic Tac Toe!</h1>
+    <div id="board">
+      <span class="squares" id="sq1"></span>
+      <span class="squares" id="sq2"></span>
+      <span class="squares" id="sq3"></span>
+      <span class="squares" id="sq4"></span>
+      <span class="squares" id="sq5"></span>
+      <span class="squares" id="sq6"></span>
+      <span class="squares" id="sq7"></span>
+      <span class="squares" id="sq8"></span>
+      <span class="squares" id="sq9"></span>
+    </div>
+  </main>
+</body>
+</html>
